Moved model classes into new package com.spindance.esp_provisioning.model
 and prevent minification of those files via proguard rules.
