+++ esp_provisioning_android/android/proguard-rules.pro
@@ -1,0 +1,5 @@
+-keepclassmembers,allowobfuscation class * {
+ @com.google.gson.annotations.SerializedName <fields>;
+}
+
+-keep class com.spindance.esp_provisioning.model.** { *; }
