+++ esp_provisioning_android/android/src/main/kotlin/com/spindance/esp_provisioning/model/AccessPoint.kt
@@ -1,4 +1,4 @@
-package com.spindance.esp_provisioning
+package com.spindance.esp_provisioning.model
 
 import com.espressif.provisioning.WiFiAccessPoint
 //import kotlinx.serialization.Serializable
