+++ esp_provisioning_android/android/src/main/kotlin/com/spindance/esp_provisioning/model/AccessPointSecurity.kt
@@ -1,4 +1,4 @@
-package com.spindance.esp_provisioning
+package com.spindance.esp_provisioning.model
 
 //import kotlinx.serialization.Serializable
