+++ esp_provisioning_android/android/src/main/kotlin/com/spindance/esp_provisioning/model/EspBleDevice.kt
@@ -1,3 +1,3 @@
-package com.spindance.esp_provisioning
+package com.spindance.esp_provisioning.model
 
 data class EspBleDevice(val name: String, val rssi: Int)
