+++ esp_provisioning_android/android/src/main/kotlin/com/spindance/esp_provisioning/model/ScannedPeripheral.kt
@@ -1,4 +1,4 @@
-package com.spindance.esp_provisioning
+package com.spindance.esp_provisioning.model
 
 import android.bluetooth.BluetoothDevice
 import android.bluetooth.le.ScanRecord
